LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/mixin-layers-expected.txt

This text should be green.
This text should be green.
This text should be green.
This text should be green.

FAIL Mixins work inside layers assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
FAIL Mixins work inside layers (anonymous) assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
FAIL Mixins in stronger layer wins assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
FAIL Mixins in stronger layer wins (source order matching layer order) assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
